make sure full request logs show URL _after_ params are interpolated
